Move post page components into app dir
