+++ app/blog/[slug]/post-content-components/post-headings.tsx
@@ -1,5 +1,5 @@
 "use client";
-import { LinkIcon } from "../../../shared/icons";
+import { LinkIcon } from "../../../../src/shared/icons";
 import toast from "react-hot-toast";
 import { RefObject, useRef } from "react";
