+++ app/blog/[slug]/post-content-components/post-image.tsx
